requirements.txt
annotated-types==0.6.0
anyio==3.7.1
asgiref==3.7.2
blinker==1.6.3
click==8.1.7
Django==4.2.6
djangorestframework==3.14.0
fastapi==0.104.0
Flask==3.0.0
gunicorn==21.2.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
packaging==23.2
pydantic==2.4.2
pydantic_core==2.10.1
pytz==2023.3.post1
sniffio==1.3.0
sqlparse==0.4.4
starlette==0.27.0
typing_extensions==4.8.0
Werkzeug==3.0.1
requests
